// car-project/src/app/models/dtoModels/carDto.ts
export interface CarDto {
  id?: string;
  maker: number;
  model: string;
  design: number;
  title: string;
  description: string;
  avaiableColors: number[];
  madeIn: number;
  cardImage: string;
  motor: number;
  transmission: number;
  seats: number;
  mainImage: string;
  images: string[];
  price: number;
}